requirements.txt
cryptography>=41.0.0
click>=8.1.0
web3>=6.0.0,<7.0.0
aiohttp>=3.9.0
uvloop>=0.19.0
pyevmasm>=0.2.3
confluent-kafka[avro]==2.3.0
fastavro>=1.8.0
eth-utils==2.1.0
black>=24.0.0
ruff>=0.4.0
mypy>=1.19.0
pydantic>=2.0.0,<3.0.0
orjson>=3.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0